// backend/src/types/productCsv.ts
export interface ProductCsv {
  Master_Category: string;
  Product_Name: string;

  Grams: string;
  Price_Per_Grams: string;
  Selling_Price: string;

  Stock: string;
  Supplier: string;

  Variation_Name_Lazada: string;
  Product_ID_Lazada: string;
  SKU_ID_Lazada: string;
  Key_Lazada: string;
  Quantity_Lazada: string;

  Product_ID_Shopee: string;
  Variation_ID_Shopee: string;
  Variation_Name_Shopee: string;

  Product_ID_Tiktok: string;
  SKU_ID_Tiktok: string;
  Variation_Name_Tiktok: string;
  Category_Tiktok: string;
  Quantity_Tiktok: string;
}